Refactorización de update (unificado en un método)
